feat: migrate config.spec
